+++ CHANGES.md
@@ -1,5 +1,6 @@
 ## v0.2.0
 * Added `--forever` option to run forever even if no child process exists
+* Added `split` filter
 
 ## v0.1.0 (2014-11-28)
 * Initial release